Dream in reality. Tallinn.
Hot weather does not deter travelers. There are a lot of tourists in Tallinn now. They try to see everything! And at the end of a hot day, enjoy a coffee drink or a snack. This is the kind of cafe I'm showing. Here, there are outdoor tables overlooking the towers .. and a cool cafe with a beautiful historical design. The cafe has a piece of the fortress wall of Tallinn.
This is so romantic. You remember historical heroes ... or everything you saw in a day. The cafe is so nice that you don't want to leave. But there are still historical mysteries ahead ... Curiosity takes over. People live in houses .. and towers nearby .. How incompatible and compatible together. Opposites attract to themselves and a kind of harmony is obtained.